renamed scp to sftp and added put/get actions
